feat(group_attendance): 출석부 UI/UX 개선 및 상세 정보 기능 추가
- 출석 현황을 한눈에 볼 수 있도록 캘린더 색상 및 요일 라벨 스타일 개선
- 날짜 클릭 시 해당 날짜의 출석 상세 정보를 보여주는 버텀 시트 추가
  - 멤버별 출석 시간, 순위, 프로필 이미지 표시
  - 일일 요약 (총 학습 시간, 참여 멤버 수, 평균 학습 시간) 카드 추가
  - 출석 데이터 로딩 및 에러 상태 처리
- 출석부 화면 로직 및 상태 관리 개선
  - 로케일 초기화 로딩 화면 추가
  - 출석 데이터와 색상 맵을 상태로 통합하여 관리
  - 코드 리팩토링 및 위젯 분리 (AttendanceBottomSheet, AttendanceSummaryCard, MemberAttendanceCard, EmptyAttendanceView)
