+++ lib/group/presentation/group_attendance/component/weekday_label.dart
@@ -1,8 +1,16 @@
+// lib/group/presentation/group_attendance/component/weekday_label.dart
+import 'package:devlink_mobile_app/core/styles/app_text_styles.dart';
 import 'package:flutter/material.dart';
 
 class WeekdayLabel extends StatelessWidget {
   final String label;
-  const WeekdayLabel({super.key, required this.label});
+  final Color color;
+
+  const WeekdayLabel({
+    super.key,
+    required this.label,
+    required this.color,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -11,10 +19,9 @@ class WeekdayLabel extends StatelessWidget {
       child: Text(
         label,
         textAlign: TextAlign.center,
-        style: const TextStyle(
-          fontSize: 16,
+        style: AppTextStyles.captionRegular.copyWith(
+          color: color,
           fontWeight: FontWeight.w600,
-          color: Color(0xFFCFCFCF),
         ),
       ),
     );
